backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pillow==9.2.0
tensorflow-cpu==2.20.0
numpy==1.23.1
